% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{pp_init_style}
\alias{pp_init_style}
\title{Setting the publication style functions}
\usage{
pp_init_style(x = c("apa"))
}
\arguments{
\item{x}{a character indicating the publication style.}
}
\description{
A list is returned which contains all the default functions for the
specified publication style.
}
\examples{
pp_opts_style$set(pp_init_style())
pp_opts_style$set(pp_init_style("apa"))

}
\seealso{
See \code{\link{pp_opts_style}} to get and set all currently used
style functions.
}

