% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\docType{data}
\name{pp_opts_out}
\alias{pp_opts_out}
\title{Output format options for the pubprint package}
\format{A list with a \code{get} and \code{set} function.}
\usage{
pp_opts_out
}
\description{
A list which functions are used to print in the correct output format
(LaTeX, HTML, Markdown or plain text). If pubprint is running inside
\code{\link[knitr]{knit}} it will automatically determine the output format
from \code{\link[knitr]{knitr}}.
}
\details{
Using \code{pp_opts_out$get()} shows all currently used output format
functions, \code{pp_opts_out$set()} allows to change them.
}
\examples{
pp_opts_out$set(pp_init_out())
pp_opts_out$set(pp_init_out("html"))

}
\seealso{
See \code{\link{pp_init_out}} for initialising this variable in the
correct way and \code{\link{pp_init_style}} for publication style.
}
\keyword{datasets}

