% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{print.pubprint}
\alias{print.pubprint}
\title{Prints a pubprint object}
\usage{
\method{print}{pubprint}(x, ...)
}
\arguments{
\item{x}{object of class \code{pubprint}.}

\item{...}{further arguments. Ignored.}
}
\description{
Prints the contents of a pubprint object
}
\details{
Prints contents of named memory and pipe of a pubprint object.
}
\examples{
ppo <- pubprint()
push(ppo) <- t.test(1:10)
print(ppo)

}

