% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{pubprint}
\alias{pubprint}
\title{Initialises a pubprint object}
\usage{
pubprint()
}
\description{
\code{pubprint} returns an empty pubprint S3 object.
}
\details{
This function initialises an empty pubprint S3 object and returns it.
This is mandatory for using the pull and push functions of the pubprint
package.
}
\examples{
ppo <- pubprint()
ppo

}
\seealso{
See \code{\link{pubprint-package}} for package documentation.
}

