% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{pull}
\alias{pull}
\title{Pulls an item from an object}
\usage{
pull(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The updated object.
}
\description{
\code{pull} is a generic function and is used to pull an item from a stack
or pipe.
}
\details{
There is no default function, so you have to see a specific \code{pull}
function for further information.
}

