% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{push<-}
\alias{push<-}
\title{Adds another item to an object}
\usage{
push(x, ...) <- value
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}

\item{value}{an item pushed to \code{x}.}
}
\value{
The updated object.
}
\description{
\code{push<-} is a generic function and is used to put another item to a stack
or pipe.
}
\details{
There is no default function, so you have to see a specific \code{push<-}
function for further information.
}

