% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{toClipboard}
\alias{toClipboard}
\title{Pastes text to clipboard}
\usage{
toClipboard(x)
}
\arguments{
\item{x}{character vector that should be pasted to clipboard.}
}
\description{
Text is written to clipboard, allowing easy pasting to other software. This
function supports only pasting of character vectors because
\code{\link[base]{writeLines}} is used. Supported operating systems are
BSD/Linux (the \code{xclip} command has to be installed), Mac OS
(\code{pbcopy} has to be installed) and Microsoft Windows.
}
\examples{
toClipboard("This is a little test.")

}

